third_party/WebKit/LayoutTests/crypto/wrapKey-unextractable-expected.txt

CONSOLE ERROR: key is not extractable
Tests that an unextractable key cannot be wrapped.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof key.extractable is 'boolean'
PASS key.extractable is false
PASS error is null
PASS successfullyParsed is true

TEST COMPLETE
